bump SDK version, restructure to keep only web
