+++ package.json
@@ -4,8 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "test:node": "ts-node node/index.ts --project node/tsconfig.json",
-    "test:web": "cd web && webpack serve --config webpack.config.js"
+    "test:web": "webpack serve --config webpack.config.js"
   },
   "devDependencies": {
     "@types/node": "^20.11.26",
@@ -20,8 +19,8 @@
     "webpack-dev-server": "^5.0.3"
   },
   "dependencies": {
-    "@heliaxdev/namada-sdk": "^0.10.0",
-    "@namada/types": "^0.2.1",
+    "@namada/sdk": "^0.12.0",
+    "@namada/types": "^0.3.0",
     "buffer": "^6.0.3",
     "typescript": "^5.6.3"
   }
