+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "moduleResolution": "node16",
-    "module": "node16"
+    "moduleResolution": "bundler",
+    "module": "esnext"
   },
-  "include": ["*.ts", "*.tsx"],
+  "include": ["./src/*.ts"],
   "exclude": ["node_modules"],
   "typeRoots": ["./node_modules/@types"]
 }
