--- web/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "bundler",
-    "module": "esnext"
-  },
-  "include": ["*.ts", "*.tsx"],
-  "exclude": ["node_modules"],
-  "typeRoots": ["./node_modules/@types"]
-}
