+++ webpack.config.js
@@ -6,7 +6,7 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 module.exports = {
   target: "web",
   mode: "development",
-  entry: path.join(__dirname, "index.ts"),
+  entry: path.join(__dirname, "src", "index.ts"),
   module: {
     rules: [
       {
@@ -28,21 +28,14 @@ module.exports = {
     alias: {
       "../package.json": path.resolve(
         __dirname,
-        "../node_modules/@heliaxdev/namada-sdk/package.json",
+        "./node_modules/@namada/sdk/package.json",
       ),
     },
   },
   devServer: {
     static: [
       path.join(__dirname, "public"),
-      path.join(
-        __dirname,
-        "..",
-        "node_modules",
-        "@heliaxdev",
-        "namada-sdk",
-        "dist",
-      ),
+      path.join(__dirname, ".", "node_modules", "@namada", "sdk", "dist"),
     ],
     compress: true,
     port: 9000,
